chore: folder restructure
