+++ app/ui/ccard/JDialog_AddCompAcc.java
@@ -1,16 +1,9 @@
-package ui.bank;
+package app.ui.ccard;
 
-/*
-		A basic implementation of the JDialog class.
-*/
+public class JDialog_AddCompAcc extends javax.swing.JDialog {
+	private CardFrm parentframe;
 
-import java.awt.*;
-import javax.swing.*;
-
-public class JDialog_AddPAcc extends javax.swing.JDialog {
-	private BankFrm parentframe;
-
-	public JDialog_AddPAcc(BankFrm parent) {
+	public JDialog_AddCompAcc(CardFrm parent) {
 		super(parent);
 		parentframe = parent;
 
@@ -20,84 +13,76 @@ public class JDialog_AddPAcc extends javax.swing.JDialog {
 		// what Visual Cafe can generate, or Visual Cafe may be unable to back
 		// parse your Java file into its visual environment.
 		// {{INIT_CONTROLS
-		setTitle("Add personal account");
+		setTitle("Add compamy account");
 		setModal(true);
 		getContentPane().setLayout(null);
-		setSize(283, 303);
+		setSize(293, 291);
 		setVisible(false);
 		JRadioButton_Chk.setText("Checkings");
 		JRadioButton_Chk.setActionCommand("Checkings");
 		getContentPane().add(JRadioButton_Chk);
-		JRadioButton_Chk.setBounds(36, 0, 84, 24);
+		JRadioButton_Chk.setBounds(36, 12, 84, 24);
 		JRadioButton_Sav.setText("Savings");
 		JRadioButton_Sav.setActionCommand("Savings");
 		getContentPane().add(JRadioButton_Sav);
-		JRadioButton_Sav.setBounds(36, 24, 84, 24);
+		JRadioButton_Sav.setBounds(36, 36, 84, 24);
 		JLabel1.setText("Name");
 		getContentPane().add(JLabel1);
 		JLabel1.setForeground(java.awt.Color.black);
-		JLabel1.setBounds(12, 84, 48, 24);
+		JLabel1.setBounds(12, 72, 48, 24);
 		JLabel2.setText("Street");
 		getContentPane().add(JLabel2);
 		JLabel2.setForeground(java.awt.Color.black);
-		JLabel2.setBounds(12, 108, 48, 24);
+		JLabel2.setBounds(12, 96, 48, 24);
 		JLabel3.setText("City");
 		getContentPane().add(JLabel3);
 		JLabel3.setForeground(java.awt.Color.black);
-		JLabel3.setBounds(12, 132, 48, 24);
+		JLabel3.setBounds(12, 120, 48, 24);
 		JLabel4.setText("State");
 		getContentPane().add(JLabel4);
 		JLabel4.setForeground(java.awt.Color.black);
-		JLabel4.setBounds(12, 156, 48, 24);
+		JLabel4.setBounds(12, 144, 48, 24);
 		JLabel5.setText("Zip");
 		getContentPane().add(JLabel5);
 		JLabel5.setForeground(java.awt.Color.black);
-		JLabel5.setBounds(12, 180, 48, 24);
-		JLabel6.setText("Birthdate");
+		JLabel5.setBounds(12, 168, 48, 24);
+		JLabel6.setText("No of employees");
 		getContentPane().add(JLabel6);
 		JLabel6.setForeground(java.awt.Color.black);
-		JLabel6.setBounds(12, 204, 96, 24);
+		JLabel6.setBounds(12, 192, 96, 24);
 		JLabel7.setText("Email");
 		getContentPane().add(JLabel7);
 		JLabel7.setForeground(java.awt.Color.black);
-		JLabel7.setBounds(12, 228, 48, 24);
+		JLabel7.setBounds(12, 216, 48, 24);
 		getContentPane().add(JTextField_NAME);
-		JTextField_NAME.setBounds(84, 84, 156, 20);
-		getContentPane().add(JTextField_CT);
-		JTextField_CT.setBounds(84, 132, 156, 20);
-		getContentPane().add(JTextField_ST);
-		JTextField_ST.setBounds(84, 156, 156, 20);
+		JTextField_NAME.setBounds(120, 72, 156, 20);
 		getContentPane().add(JTextField_STR);
-		JTextField_STR.setBounds(84, 108, 156, 20);
+		JTextField_STR.setBounds(120, 96, 156, 20);
+		getContentPane().add(JTextField_CT);
+		JTextField_CT.setBounds(120, 120, 156, 20);
+		getContentPane().add(JTextField_ST);
+		JTextField_ST.setBounds(120, 144, 156, 20);
+		getContentPane().add(JTextField_STR);
 		getContentPane().add(JTextField_ZIP);
-		JTextField_ZIP.setBounds(84, 180, 156, 20);
-		getContentPane().add(JTextField_BD);
-		JTextField_BD.setBounds(84, 204, 156, 20);
+		JTextField_ZIP.setBounds(120, 168, 156, 20);
+		getContentPane().add(JTextField_NoOfEmp);
+		JTextField_NoOfEmp.setBounds(120, 192, 156, 20);
 		getContentPane().add(JTextField_EM);
-		JTextField_EM.setBounds(84, 228, 156, 20);
+		JTextField_EM.setBounds(120, 216, 156, 20);
 		JButton_OK.setText("OK");
 		JButton_OK.setActionCommand("OK");
 		getContentPane().add(JButton_OK);
-		JButton_OK.setBounds(48, 264, 84, 24);
-		JButton_Cancel.setText("Cancel");
-		JButton_Cancel.setActionCommand("Cancel");
-		getContentPane().add(JButton_Cancel);
-		JButton_Cancel.setBounds(156, 264, 84, 24);
-		getContentPane().add(JTextField_ACNR);
-		JTextField_ACNR.setBounds(84, 60, 156, 20);
-		JLabel8.setText("Acc Nr");
-		getContentPane().add(JLabel8);
-		JLabel8.setForeground(java.awt.Color.black);
-		JLabel8.setBounds(12, 60, 48, 24);
+		JButton_OK.setBounds(48, 252, 84, 24);
+		JButton_Calcel.setText("Cancel");
+		JButton_Calcel.setActionCommand("Cancel");
+		getContentPane().add(JButton_Calcel);
+		JButton_Calcel.setBounds(156, 252, 84, 24);
 		// }}
 
 		// {{REGISTER_LISTENERS
-		SymMouse aSymMouse = new SymMouse();
-		JRadioButton_Chk.addMouseListener(aSymMouse);
-		JRadioButton_Sav.addMouseListener(aSymMouse);
 		SymAction lSymAction = new SymAction();
 		JButton_OK.addActionListener(lSymAction);
-		JButton_Cancel.addActionListener(lSymAction);
+		JButton_Calcel.addActionListener(lSymAction);
 		// }}
 	}
 
@@ -116,51 +101,23 @@ public class JDialog_AddPAcc extends javax.swing.JDialog {
 	javax.swing.JTextField JTextField_ST = new javax.swing.JTextField();
 	javax.swing.JTextField JTextField_STR = new javax.swing.JTextField();
 	javax.swing.JTextField JTextField_ZIP = new javax.swing.JTextField();
-	javax.swing.JTextField JTextField_BD = new javax.swing.JTextField();
+	javax.swing.JTextField JTextField_NoOfEmp = new javax.swing.JTextField();
 	javax.swing.JTextField JTextField_EM = new javax.swing.JTextField();
 	javax.swing.JButton JButton_OK = new javax.swing.JButton();
-	javax.swing.JButton JButton_Cancel = new javax.swing.JButton();
-	javax.swing.JTextField JTextField_ACNR = new javax.swing.JTextField();
-	javax.swing.JLabel JLabel8 = new javax.swing.JLabel();
+	javax.swing.JButton JButton_Calcel = new javax.swing.JButton();
 	// }}
-
-	class SymMouse extends java.awt.event.MouseAdapter {
-		public void mouseClicked(java.awt.event.MouseEvent event) {
-			Object object = event.getSource();
-			if (object == JRadioButton_Chk)
-				JRadioButtonChk_mouseClicked(event);
-			else if (object == JRadioButton_Sav)
-				JRadioButtonSav_mouseClicked(event);
-		}
-	}
-
-	void JRadioButtonChk_mouseClicked(java.awt.event.MouseEvent event) {
-		// When Checking radio is clicked make this radio on
-		// and make Saving account radio off
-		JRadioButton_Chk.setSelected(true);
-		JRadioButton_Sav.setSelected(false);
-	}
-
-	void JRadioButtonSav_mouseClicked(java.awt.event.MouseEvent event) {
-		// When Saving radio is clicked make this radio on
-		// and make Checking account radio off
-		JRadioButton_Chk.setSelected(false);
-		JRadioButton_Sav.setSelected(true);
-
-	}
 
 	class SymAction implements java.awt.event.ActionListener {
 		public void actionPerformed(java.awt.event.ActionEvent event) {
 			Object object = event.getSource();
 			if (object == JButton_OK)
 				JButtonOK_actionPerformed(event);
-			else if (object == JButton_Cancel)
+			else if (object == JButton_Calcel)
 				JButtonCalcel_actionPerformed(event);
 		}
 	}
 
 	void JButtonOK_actionPerformed(java.awt.event.ActionEvent event) {
-		parentframe.accountnr = JTextField_ACNR.getText();
 		parentframe.clientName = JTextField_NAME.getText();
 		parentframe.street = JTextField_STR.getText();
 		parentframe.city = JTextField_CT.getText();
@@ -172,10 +129,11 @@ public class JDialog_AddPAcc extends javax.swing.JDialog {
 			parentframe.accountType = "S";
 		parentframe.newaccount = true;
 		dispose();
+
 	}
 
 	void JButtonCalcel_actionPerformed(java.awt.event.ActionEvent event) {
-		// make this frame invisible if Cancel button is clicked
 		dispose();
+
 	}
 }
